Extract magic leap math helpers to ml-math.{h,cc}
