+++ deps/exokit-bindings/magicleap/include/magicleap.h
@@ -19,6 +19,7 @@
 #include <condition_variable>
 #include <algorithm>
 #include <webgl.h>
+#include <magicleap/include/ml-math.h>
 #include <egl.h>
 #include <ml_graphics.h>
 #include <ml_head_tracking.h>
